+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 0.0.7
+
+* optional bracket for single positional argument
+* dot (.) separated paths can be defined as Map e.g. Alignment in README.md
+
+## 0.0.6
+
+* sdk version requirement changed
+
+## 0.0.5
+
+* Example updated
+
 ## 0.0.4
 
 * Example updated